Use Kibo UI theme switcher
