+++ app/layout.tsx
@@ -1,23 +1,23 @@
 import { mono, sans } from '@/lib/fonts';
 import './globals.css';
-import { Footer } from '@/components/footer';
-import { Header } from '@/components/header';
-import { JsonLd } from '@/components/json-ld';
-import { Navigation } from '@/components/navigation';
-import { ThemeSwitcher } from '@/components/theme-switcher';
-import { WindowsEmojiPolyfill } from '@/components/windows-emoji-polyfill';
-import { cn } from '@/lib/utils';
 import { Analytics } from '@vercel/analytics/react';
 import { ThemeProvider } from 'next-themes';
 import type { ReactNode } from 'react';
 import { Toaster } from 'sonner';
+import { Footer } from '@/components/footer';
+import { Header } from '@/components/header';
+import { JsonLd } from '@/components/json-ld';
+import { Navigation } from '@/components/navigation';
+import { Theme } from '@/components/theme';
+import { WindowsEmojiPolyfill } from '@/components/windows-emoji-polyfill';
+import { cn } from '@/lib/utils';
 
 type RootLayoutProps = {
   children: ReactNode;
 };
 
 const RootLayout = ({ children }: RootLayoutProps) => (
-  <html lang="en" suppressHydrationWarning className="scroll-smooth">
+  <html className="scroll-smooth" lang="en" suppressHydrationWarning>
     <body
       className={cn(
         sans.variable,
@@ -28,8 +28,8 @@ const RootLayout = ({ children }: RootLayoutProps) => (
       <ThemeProvider
         attribute="class"
         defaultTheme="system"
-        enableSystem
         disableTransitionOnChange
+        enableSystem
       >
         <Header />
         <div className="grid grid-cols-[200px_1fr]">
@@ -42,7 +42,6 @@ const RootLayout = ({ children }: RootLayoutProps) => (
           </div>
         </div>
         <Toaster />
-        <ThemeSwitcher />
         <WindowsEmojiPolyfill />
         <JsonLd />
         <Analytics />
