+++ components/navigation.tsx
@@ -1,8 +1,9 @@
 'use client';
 
+import { usePathname } from 'next/navigation';
 import { Link } from '@/components/link';
 import { cn } from '@/lib/utils';
-import { usePathname } from 'next/navigation';
+import { Theme } from './theme';
 
 const links = [
   {
@@ -46,20 +47,25 @@ export const Navigation = () => {
   const pathname = usePathname();
 
   return (
-    <ul className="sticky top-16 flex h-[calc(100vh-4rem)] flex-col justify-center p-4">
-      {links.map(({ href, label, active }) => (
-        <li key={href}>
-          <Link
-            href={href}
-            className={cn(
-              'block rounded-md px-3 py-2 text-sm hover:bg-muted',
-              active(pathname) ? 'text-primary' : 'border-none'
-            )}
-          >
-            {label}
-          </Link>
-        </li>
-      ))}
-    </ul>
+    <div className='sticky top-16 flex h-[calc(100vh-4rem)] flex-col justify-center p-4'>
+      <ul>
+        {links.map(({ href, label, active }) => (
+          <li key={href}>
+            <Link
+              className={cn(
+                'block rounded-md px-3 py-2 text-sm hover:bg-muted',
+                active(pathname) ? 'text-primary' : 'border-none'
+              )}
+              href={href}
+            >
+              {label}
+            </Link>
+          </li>
+        ))}
+      </ul>
+      <div className="absolute bottom-0 left-0 p-4">
+        <Theme />
+      </div>
+    </div>
   );
 };
