+++ content/pages/projects.mdx
@@ -19,4 +19,4 @@ I'm also working on [Kibo UI](https://www.kibo-ui.com/), a custom registry of co
 
 <Video url="https://www.youtube.com/shorts/1yoW9OH9YAk" controls />
 
-I also made some open source alternatives to existing tools, including [Tersa](https://www.tersa.ai/), an open source alternative to Flora, and [Eververse](https://www.eververse.ai/), an open source alternative to Productboard.
+I maintain some open source alternatives to existing tools, including [Tersa](https://www.tersa.ai/), an open source alternative to Flora, and [Eververse](https://www.eververse.ai/), an open source alternative to Productboard.
