+++ package.json
@@ -7,6 +7,7 @@
     "build": "next build --turbopack"
   },
   "dependencies": {
+    "@radix-ui/react-use-controllable-state": "^1.2.2",
     "@t3-oss/env-nextjs": "^0.13.8",
     "@vercel/analytics": "^1.5.0",
     "class-variance-authority": "^0.7.1",
